Alter extendWidgetControl to a (mostly) pure function
Changes the function to accept an object, alter that object's prototype,
and return it. The caller is then responsible for passing the
constructor function to change.

This allows the function to be more easily tested since it does not
directly mutate global variables (although of course changing the
prototype of a global variable will have a a global effect).

Also adds `wp` and `_` as parameters to the IIFE to make their usage
explicit.